refactor: extract the dashboard/presentation layer into dashboard.py (step 4)
Final step of the split. Move 15 presentation functions (issue charts, the
dashboard notebook tab + search/copy, progress-bar styling, and the
rate-limit / thread-status displays) into dashboard.py. generate.py drops
from 2,564 to 2,013 lines — the original 6,802-line monolith is now an
~2,000-line app shell plus four focused modules.

The coupling here is to GUI-widget handles created in generate.py's
main-window construction, so:
  - 10 widget handles (dashboard_notebook, status_bar, the *_label stat
    labels, budget/cost/token labels) move to app_state as None placeholders
    and are qualified to app_state.X (construction + the moved functions).
  - the globals()-based per-API slot-label registry becomes
    app_state.slot_widgets (a dict): construction writes
    app_state.slot_widgets[name] = label, update_dashboard reads it. A bare
    globals()[name] would resolve against dashboard.py's namespace after the
    move and break.
  - the stale `'threads' in globals()` guard (left over from step 2, where
    threads moved to app_state, so it was always False) is simplified to
    `app_state.threads`, restoring the spawned-thread count.

dashboard.py imports app_state + the helper modules and does NOT import
generate.py (dashboard->staying crossings are zero). generate.py pulls back
the entrypoints construction/lifecycle call via `from dashboard import ...`.

Verified: py_compile on all modules; pyflakes shows no undefined names in
generate.py or dashboard.py; and a headless xvfb run that builds the full
app, confirms the widget placeholders + slot_widgets registry are populated,
that dashboard.py has no generate dependency, that update_dashboard() and
update_thread_status_display() actually run post-move, and that the engine
and ConfigEditor remain wired.
